pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "boto3>=1.40.25",
    "botocore>=1.40.25",
    "pyrogram>=2.0.106",
    "python-dotenv>=1.1.1",
    "tgcrypto>=1.2.5",
]
